# docs/api/openapi.bundled.yaml
openapi: 3.1.0
info:
  title: RoboScope API
  version: 0.1.1
  license:
    name: MIT
    identifier: MIT
  summary: Платформа экспресс-оценки роботизации объекта (ЛЦТ 2026, задача №1, ФЦ БАС)
  description: |
    Контракт между фронтендом и бэкендом. Источник истины по API: бэкенд обязан ему соответствовать,
    фронтенд разрабатывается по нему с mock-сервером (см. README.md рядом).

    **Конвенции**
    - Префикс `/api/v1` входит в путь (без `servers`), чтобы mock, vite-proxy и vercel rewrite совпадали.
    - Авторизация: `Authorization: Bearer <access>`. Для `X-Client: web` refresh-токен живёт в httpOnly-cookie,
      обновление требует `X-CSRF-Token`. Гостевые эндпоинты помечены `security: []`.
    - Ошибки: всегда `Problem` с `error_code` из закрытого списка.
    - Деньги — в рублях (`amount_rub`), не в миллионах. Время — ISO 8601 UTC.
    - Любое значение, влияющее на расчёт, несёт `provenance` (статус и источник).
    - Долгие операции (симуляция, Монте-Карло, отчёты, обогащение) возвращают 202 + `Job`/`SimulationRun`;
      прогресс — SSE `GET /api/v1/jobs/{job_id}/stream` или опрос.
    - Пагинация: `page` (с 1) и `page_size`; ответ содержит `items`, `total`, `page`, `page_size`.
    - Версии входов (`VersionStamp`) сохраняются с каждым расчётом — расчёт воспроизводим (ТЗ 3.1.5).
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/jobs/{job_id}:
    get:
      tags:
        - system
      operationId: getJob
      summary: Состояние фоновой задачи
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/object-types/{object_type}:
    get:
      tags:
        - reference
      operationId: getObjectType
      summary: Тип объекта — схема формы параметров
      security: []
      parameters:
        - name: object_type
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ObjectTypeKey'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectType'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/solution-types:
    get:
      tags:
        - reference
      operationId: listSolutionTypes
      summary: Типы роботизированных решений
      security: []
      parameters:
        - name: object_type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ObjectTypeKey'
        - name: process_key
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SolutionType'
...
  /api/v1/norms:
    get:
      tags:
        - reference
      operationId: listNorms
      summary: Реестр нормативов расчётной модели с источниками
      description: Показывается пользователю на экране «Методика» и в отчёте. Каждый норматив имеет источник и обоснование.
      security: []
      parameters:
        - name: version
          in: query
          required: false
          schema:
            type: string
          description: Версия набора; по умолчанию текущая
        - name: category
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/NormCategory'
        - name: object_type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ObjectTypeKey'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - version
                  - items
                properties:
                  version:
                    type: string
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Norm'
...
  /api/v1/catalog/products:
    get:
      tags:
        - catalog
      operationId: listProducts
      summary: Каталог продуктов с фильтрами, поиском и сортировкой (ТЗ 3.3.7)
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: q
          in: query
          schema:
            type: string
          description: Полнотекстовый или семантический поиск
        - name: object_type
          in: query
          schema:
            $ref: '#/components/schemas/ObjectTypeKey'
        - name: process_key
          in: query
          schema:
            type: string
        - name: solution_type
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
        - name: industry
          in: query
          schema:
            type: string
        - name: manufacturer_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ProductStatus'
          style: form
          explode: false
        - name: badge
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Badge'
          style: form
          explode: false
        - name: price_min_rub
          in: query
          schema:
            type: number
        - name: price_max_rub
          in: query
          schema:
            type: number
        - name: payload_min_kg
          in: query
          schema:
            type: number
        - name: min_completeness
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 1
        - name: sort
          in: query
          schema:
            type: string
            enum:
              - relevance
              - name
              - price_asc
              - price_desc
              - completeness
              - trl
              - updated
            default: relevance
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'
  /api/v1/catalog/facets:
    get:
      tags:
        - catalog
      operationId: getCatalogFacets
      summary: Значения фильтров с количествами
      security: []
      parameters:
        - name: object_type
          in: query
          schema:
            $ref: '#/components/schemas/ObjectTypeKey'
        - name: process_key
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogFacets'
...
  /api/v1/catalog/products/{product_id}:
    get:
      tags:
        - catalog
      operationId: getProduct
      summary: Карточка продукта — предложения, ТТХ с источниками, кейсы
      security: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/catalog/compare:
    post:
      tags:
        - catalog
      operationId: compareProducts
      summary: Сравнение 2–5 продуктов по группам характеристик
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompareRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompareResult'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /api/v1/admin/catalog/products:
    post:
      tags:
        - admin
      operationId: adminCreateProduct
      summary: Добавить продукт вручную (ТЗ 3.3.5)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductWrite'
      responses:
        '201':
          description: Создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/catalog/products/{product_id}:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    patch:
      tags:
        - admin
      operationId: adminUpdateProduct
      summary: Редактировать продукт
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductWrite'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
    delete:
      tags:
        - admin
      operationId: adminDeleteProduct
      summary: Скрыть продукт (мягкое удаление — расчёты, где он использован, сохраняются)
      responses:
        '204':
          description: Скрыт
...
  /api/v1/integrations/catalog/products:
    post:
      tags:
        - integrations
      operationId: integrationUpsertProducts
      summary: Обновить продукты каталога из внешнего источника по ключу API (для вендоров и ФЦ БАС)
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - items
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/ProductWrite'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogImportResult'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access-токен из /auth/login или /auth/refresh
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Ключ интеграции из /me/api-keys (только эндпоинты /integrations/*)
  schemas:
...
    Job:
      type: object
      description: |
        Фоновая задача. Долгие операции (симуляция, Монте-Карло, отчёт, обогащение) возвращают Job сразу;
        результат забирается по result_url или по SSE-потоку прогресса.
      required:
        - id
        - type
        - status
        - progress
        - created_at
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/JobType'
        status:
          $ref: '#/components/schemas/JobStatus'
        progress:
          type: number
          minimum: 0
          maximum: 1
        stage:
          type:
            - string
            - 'null'
          example: Прогон пикового дня, час 14 из 24
        created_at:
          type: string
          format: date-time
        started_at:
          type:
            - string
            - 'null'
          format: date-time
        finished_at:
          type:
            - string
            - 'null'
          format: date-time
        result_url:
          type:
            - string
            - 'null'
          description: Куда идти за результатом, когда status=done
        error:
          oneOf:
            - $ref: '#/components/schemas/Problem'
            - type: 'null'
...
    ObjectType:
      type: object
      required:
        - key
        - name
        - industry
        - depth
        - parameter_groups
        - processes
      properties:
        key:
          $ref: '#/components/schemas/ObjectTypeKey'
        name:
          type: string
          example: Склад
        description:
          type: string
        industry:
          type: string
          example: Торговля и услуги
        depth:
          type: string
          description: full — расчёт, симуляция, отчёт; basic — параметры, подбор, аналитическая экономика без симуляции.
          enum:
            - full
            - basic
        demo_projects:
          type: array
          items:
            type: object
            required:
              - key
              - name
            properties:
              key:
                type: string
                example: warehouse_demo_01
              name:
                type: string
                example: Склад 20 000 м², 180 сотрудников (данные организатора)
              description:
                type: string
        parameter_groups:
          type: array
          items:
            $ref: '#/components/schemas/ParameterGroup'
        processes:
          type: array
          items:
            $ref: '#/components/schemas/ProcessDef'
        layout_templates:
          type: array
          items:
            type: string
            example: warehouse_cross_dock
...
    Norm:
      type: object
      description: |
        Норматив расчётной модели. ТЗ 3.5.1 запрещает недокументированные коэффициенты — каждый норматив имеет источник
        и обоснование, и пользователь видит его в интерфейсе (ТЗ 3.2.5, 3.5.8).
      required:
        - key
        - name
        - value
        - unit
        - category
        - source
        - editable_by_user
      properties:
        key:
          type: string
          example: amr_utilization_target
        name:
          type: string
          example: Целевая загрузка AMR
        value:
          type: number
          example: 0.8
        unit:
          type: string
          example: доля
        category:
          $ref: '#/components/schemas/NormCategory'
        range:
          type: object
          properties:
            min:
              type: number
              example: 0.7
            max:
              type: number
              example: 0.85
        source:
          $ref: '#/components/schemas/Source'
        rationale:
          type: string
          example: 'Лист «Легенда» датасета организатора: загрузка AMR 70–85 %'
        affects:
          type: array
          items:
            type: string
          description: Ключи метрик, на которые влияет
        editable_by_user:
          type: boolean
          description: Можно ли переопределить в сценарии (ТЗ 3.5.3)
        object_types:
          type: array
          items:
            $ref: '#/components/schemas/ObjectTypeKey'
...
    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/Product'
        - type: object
          required:
            - offers
            - specs
            - cases
            - sources
          properties:
            description:
              type:
                - string
                - 'null'
            offers:
              type: array
              items:
                $ref: '#/components/schemas/ProductOffer'
            specs:
              type: array
              description: Все характеристики, сгруппированы по group на фронте.
              items:
                $ref: '#/components/schemas/Spec'
            cases:
              type: array
              items:
                $ref: '#/components/schemas/ProductCase'
            sources:
              type: array
              items:
                $ref: '#/components/schemas/Source'
            integration_notes:
              type:
                - string
                - 'null'
              example: Интеграция с WMS по REST API вендора; QR-разметка пола
            missing_key_specs:
              type: array
              description: Ключевые ТТХ без данных — влияют на статус «требует проверки».
              items:
                type: string
            similar_products:
              type: array
              items:
                $ref: '#/components/schemas/Product'
...
    JobType:
      type: string
      enum:
        - simulation
        - monte_carlo
        - report
        - enrichment
        - import
        - calibration
    JobStatus:
      type: string
      enum:
        - queued
        - running
        - done
        - failed
        - cancelled
...
    ObjectTypeKey:
      type: string
      description: warehouse — склад (в полной глубине), airport, hospital — параметры и подбор; custom — свой объект.
      enum:
        - warehouse
        - airport
        - hospital
        - custom
...
    SizingModel:
      type: string
      description: Каким расчётом определяется производительность робота для процесса.
      enum:
        - transport_cycle
        - goods_to_person
        - area_coverage
        - station_throughput
        - tow_train
        - elevator_cycle
    SolutionType:
      type: object
      required:
        - key
        - name
        - sizing_model
        - capability_keys
      properties:
        key:
          type: string
          example: amr_transport
        name:
          type: string
          example: AMR — транспортный мобильный робот
        description:
          type: string
        sizing_model:
          description: null — категория вне моделей расчёта (дроны, морские роботы и т.п.), только каталог.
          oneOf:
            - $ref: '#/components/schemas/SizingModel'
            - type: 'null'
        capability_keys:
          type: array
          description: Ключи ТТХ, участвующие в жёстких проверках для этого типа.
          items:
            type: string
            example: payload_kg
        processes:
          type: array
          items:
            type: string
        products_count:
          type: integer
...
    NormCategory:
      type: string
      enum:
        - capex
        - opex
        - labor
        - finance
        - operations
        - simulation
        - sizing
        - risk
...
    ProductStatus:
      type: string
      description: Стадия из каталога организатора.
      enum:
        - operation
        - piloting
        - rnd
    Badge:
      type: string
      description: |
        in_registry_719 — есть в реестре промпродукции (ПП 719); tested_fcbas — протестировано ФЦ БАС;
        specs_confirmed — все ключевые ТТХ подтверждены первичными источниками; domestic — отечественный производитель;
        has_cases — есть реализованные внедрения.
      enum:
        - in_registry_719
        - tested_fcbas
        - specs_confirmed
        - domestic
        - has_cases
...
    ProductList:
      type: object
      required:
        - items
        - page
        - page_size
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
...
    CatalogFacets:
      type: object
      description: Значения фильтров с количествами — для панели фильтров каталога.
      properties:
        solution_types:
          type: array
          items:
            $ref: '#/components/schemas/Facet'
        manufacturers:
          type: array
          items:
            $ref: '#/components/schemas/Facet'
        statuses:
          type: array
          items:
            $ref: '#/components/schemas/Facet'
        industries:
          type: array
          items:
            $ref: '#/components/schemas/Facet'
        badges:
          type: array
          items:
            $ref: '#/components/schemas/Facet'
        object_types:
          type: array
          items:
            $ref: '#/components/schemas/Facet'
        price_range:
          type: object
          properties:
            min_rub:
              type: number
            max_rub:
              type: number
...
    CompareRequest:
      type: object
      required:
        - product_ids
      properties:
        product_ids:
          type: array
          minItems: 2
          maxItems: 5
          items:
            type: string
            format: uuid
        object_type:
          oneOf:
            - $ref: '#/components/schemas/ObjectTypeKey'
            - type: 'null'
        process_key:
          type:
            - string
            - 'null'
          description: Если задан — в таблицу добавляются оценки для процесса
        project_id:
          type:
            - string
            - 'null'
          format: uuid
          description: Если задан — проверки совместимости с объектом
...
    CompareResult:
      type: object
      required:
        - products
        - rows
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        rows:
          type: array
          items:
            type: object
            required:
              - spec_key
              - name
              - group
              - values
            properties:
              spec_key:
                type: string
              name:
                type: string
              group:
                $ref: '#/components/schemas/SpecGroup'
              unit:
                type:
                  - string
                  - 'null'
              better:
                type: string
                enum:
                  - higher
                  - lower
                  - none
              values:
                type: object
                description: product_id → Spec
                additionalProperties:
                  $ref: '#/components/schemas/Spec'
              best_product_id:
                type:
                  - string
                  - 'null'
                format: uuid
        compatibility:
          type: object
          description: product_id → результат жёстких проверок относительно проекта (если передан project_id).
          additionalProperties:
            $ref: '#/components/schemas/CandidateStatus'
...
    ProductWrite:
      type: object
      required:
        - name
        - manufacturer_name
        - solution_type
        - status
      properties:
        name:
          type: string
        manufacturer_name:
          type: string
        manufacturer_country:
          type: string
          default: RU
        manufacturer_region:
          type:
            - string
            - 'null'
        solution_type:
          type: string
        subtype:
          type:
            - string
            - 'null'
        status:
          $ref: '#/components/schemas/ProductStatus'
        trl:
          type:
            - integer
            - 'null'
        market_potential:
          type:
            - number
            - 'null'
        description:
          type:
            - string
            - 'null'
        image_url:
          type:
            - string
            - 'null'
        object_types:
          type: array
          items:
            $ref: '#/components/schemas/ObjectTypeKey'
        processes:
          type: array
          items:
            type: string
        badges:
          type: array
          items:
            $ref: '#/components/schemas/Badge'
        offers:
          type: array
          items:
            type: object
            required:
              - industry
              - scenario
              - price
            properties:
              industry:
                type: string
              scenario:
                type: string
              price:
                $ref: '#/components/schemas/Money'
              cases_text:
                type:
                  - string
                  - 'null'
              source:
                $ref: '#/components/schemas/SourceWrite'
...
  parameters:
    JobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
    ProductId:
      name: product_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ValidationError:
      description: Ошибка валидации запроса
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Forbidden:
      description: Недостаточно прав или чужой проект
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...